Change id types to BIGINT & BIGSERIAL

+++ assets/migrations/scripts/20200319093703_create_table_client_form_metadata.sql
@@ -18,7 +18,7 @@
 -- Migration SQL that makes the change goes here.
 CREATE TABLE core.client_form_metadata
 (
-    id INTEGER REFERENCES core.client_form(id),
+    id BIGINT REFERENCES core.client_form(id),
     identifier character varying NOT NULL,
     jurisdiction character varying,
     version character varying NOT NULL,
